refactor(api): split api.py into api/ module (A1)
Splits the 1373-line monolithic api.py into six focused submodules:
- api/_helpers.py: _get_manager, _validate_oauth_state, _get_setup_client
- api/setup.py: all setup-wizard views + OAuth callback (7 classes)
- api/refresh.py: RefreshTriggerView, RefreshStatusView
- api/data.py: BalancesView, TransactionsView, SummaryView, TransferChainsView
- api/static.py: StaticView (LRU-cached file serving)
- api/demo.py: DemoToggleView, DemoDataView
- api/__init__.py: async_register_api + full re-exports for backwards compat

The import 'from .api import async_register_api' in __init__.py continues
to work unchanged. Updates test patch target to api.refresh._get_manager.
